main - Trying to mock a bit differently
The CGI program stdin is a text interface that is likely decoding using an incorrect locale due to an old bug. I could "fix" it
by setting the encoding in the HTTPD service, but that's a lot of config I'm not yet sure about. This change ceates a dummy mock
object that has the buffer attribute I need to access.